backend/requirements.txt
# DataPilot backend – FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.0

# BigQuery (POC data & later agent queries)
google-cloud-bigquery>=3.14.0

# Gemini LLM
langchain-core>=0.3.0 # engine for our lanchain
langchain-google-genai>=2.0.0 # Allows langchain to talk to Gemini

# Agent orchestration
langgraph>=0.2.0

# Supabase (auth + chat persistence)
supabase>=2.0.0

# PostgreSQL (optional, for DATABASE_TYPE=postgres)
psycopg2-binary>=2.9.0

# Encrypted user data source configs
cryptography>=42.0.0
